Converted PlantHolderVisualizer to component and system.
