+++ Content.Client/Botany/PlantHolderVisualsComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Client.Botany;
+
+[RegisterComponent]
+public sealed class PlantHolderVisualsComponent : Component
+{
+}
